Created CaptureLine and improved ExecLine
I made a simplified version of ExecLine for use in just dumping output
of a command, without all of the commenting and time-stamping.  In order
to get around formatting issues, I endabled paste mode.

I additionally enabled paste mode for ExecLine and disabled the redraw
function, which means that the entire output can be removed with a
single undo now.